move migration code to separate files
